Исправление критических ошибок компиляции
- HandRank: name -> displayName (конфликт с Enum.name)
- RangeAnalyzer: добавлен .toDouble() для evaluatePreFlopStrength
- Theme: исправлено использование WindowInsetsControllerCompat
- Все ошибки компиляции устранены

+++ app/src/main/java/com/holdem/poker/strategy/RangeAnalyzer.kt
@@ -84,7 +84,7 @@ class RangeAnalyzer {
                 )
                 evaluation.rank.value / 10.0
             } else {
-                evaluatePreFlopStrength(hand.first, hand.second)
+                evaluatePreFlopStrength(hand.first, hand.second).toDouble()
             }
         }
